Survival analysis. Event: becoming a parent.
For this example we are using data from the Chilean VIII National Youth Survey, focusing on the timing of parenthood onset, a critical life event with significant implications for individual trajectories and policy interventions. The objective is to understand how factors like gender, socioeconomic background (SES), and religious identification influence the likelihood and timing of the event: becoming a parent.